+++ NEWS.md
@@ -9,6 +9,12 @@
 * Replace `readLines()` with `brio::read_lines()` to make parsers robust to 
   Windows encoding issues (similar to https://stackoverflow.com/questions/18789330/r-on-windows-character-encoding-hell).
 
+* Use `"sf"` as default download method for `download_boundaries(..., method = "sf")`.
+  Add arguments `quiet_download` and `quiet_parse = TRUE` to 
+  `download_boundaries()`. `quiet_download` (default `FALSE`) controls `download.file()` 
+  messages. `quiet_parse` (default `TRUE`) controls messages from `sf::st_read()` when
+  `method = "sf"`.
+
 ## rdhs 0.7.0
 
 * Add CITATION info.